fix(ci-ronda1): correcciones de compilación C++/C# detectadas por el runner
Núcleo: RTL_OSVERSIONINFOEXW (wProductType/SP), WndProc público,
ternario inválido en párrafo-alineación, DWRITE_FONT_STRETCH_NORMAL en
CreateTextFormat, pinceles ID2D1SolidColorBrush reales para
DrawTextW/FillRectangle, StringFormat sin ctor-copia protegido, ToWide
eliminado sobre wstring, NOMINMAX sin duplicar, variables sin uso.

C#: Add(Make(string)) en overlay ipc, aritmética u32 del key schedule
Twofish (2*i*RHO con desbordamiento), casts uint en S-boxes, CS0675 del
varint, WindowsBase referenciado en Lite (System.IO.Packaging).

+++ src/managed/FusionShared/Bible/SQLiteFileReader.cs
@@ -245,7 +245,7 @@ namespace Fusion.Shared.Bible
             }
             if (pos < b.Length)
             {
-                v = (v << 8) | b[pos++];
+                v = (v << 8) | (ulong)b[pos++];
             }
             return (long)v;
         }
